% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/electrical_conductivity_bulk_model.R
\name{bulk_ec_Archie}
\alias{bulk_ec_Archie}
\title{Bulk electrical conductivity of fluid-solid mixtures based on Archie's law}
\usage{
bulk_ec_Archie(phi, sigma_f, m, a = 1)
}
\arguments{
\item{phi}{porosity [-]}

\item{sigma_f}{fluid electrical conductivity [S/m]}

\item{m}{cementation exponent [-]}

\item{a}{tortuosity factor [-]}
}
\value{
bulk electrical conductivity [S/m]
}
\description{
Bulk electrical conductivity of fluid-solid mixtures based on Archie's law
}
