% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/electrical_conductivity_bulk_model.R
\name{bulk_ec_HSupper}
\alias{bulk_ec_HSupper}
\title{Bulk electrical conductivity of fluid-solid mixtures from Hashin-Shtrikman upper bound}
\usage{
bulk_ec_HSupper(phi, sigma_f, sigma_s)
}
\arguments{
\item{phi}{porosity [-]}

\item{sigma_f}{fluid electrical conductivity [S/m]}

\item{sigma_s}{Solid electrical conductivity [S/m]}
}
\value{
bulk electrical conductivity [S/m]
}
\description{
Bulk electrical conductivity of fluid-solid mixtures from Hashin-Shtrikman upper bound
}
