% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/electrical_conductivity_bulk_model.R
\name{bulk_ec_ModifiedArchie}
\alias{bulk_ec_ModifiedArchie}
\title{Bulk electrical conductivity of fluid-solid mixtures based on modified Archie's law}
\usage{
bulk_ec_ModifiedArchie(phi, sigma_f, sigma_s, m)
}
\arguments{
\item{phi}{porosity [-]}

\item{sigma_f}{fluid electrical conductivity [S/m]}

\item{sigma_s}{Solid electrical conductivity [S/m]}

\item{m}{cementation exponent [-]}
}
\value{
bulk electrical conductivity [S/m]
}
\description{
Bulk electrical conductivity of fluid-solid mixtures based on modified Archie's law
}
