% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/driesner07_util.R
\name{driesner07_H2O_NaCl_estimate_density_from_M0}
\alias{driesner07_H2O_NaCl_estimate_density_from_M0}
\title{Estimate density from concentration}
\usage{
driesner07_H2O_NaCl_estimate_density_from_M0(M0)
}
\arguments{
\item{M0}{Molar concentration at room temperature [mol/L]}
}
\value{
density [kg/m3]
}
\description{
Estimate density from concentration
}
