% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/driesner07_util.R
\name{driesner07_H2O_NaCl_get_phase_relation_on_Tp_space}
\alias{driesner07_H2O_NaCl_get_phase_relation_on_Tp_space}
\title{get information about H2O-NaCl phase relation at the given NaCl mass fraction, based on Driesner & Heinrich (2007)}
\usage{
driesner07_H2O_NaCl_get_phase_relation_on_Tp_space(massfrac_NaCl, TC_max = 800)
}
\arguments{
\item{massfrac_NaCl}{NaCl mass fraction (bulk) [kg/kg]}

\item{TC_max}{Maximum temperature to be calculated [K]}
}
\value{
a list object including all the phase relation information
}
\description{
get information about H2O-NaCl phase relation at the given NaCl mass fraction, based on Driesner & Heinrich (2007)
}
