% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/driesner07_util.R
\name{driesner07_H2O_NaCl_get_properties_TpX}
\alias{driesner07_H2O_NaCl_get_properties_TpX}
\title{Get H2O-NaCl fluid properties at the given temperature, pressure, and NaCl mass fraction}
\usage{
driesner07_H2O_NaCl_get_properties_TpX(TK, p, X, print = FALSE)
}
\arguments{
\item{TK}{Temeprature [K]}

\item{p}{Pressure [Pa]}

\item{X}{Bulk NaCl mass fraction [kg/kg]}

\item{print}{TRUE for printing the calculated properties}
}
\value{
a nested list object including the fluid properties
}
\description{
Get H2O-NaCl fluid properties at the given temperature, pressure, and NaCl mass fraction
}
