% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/driesner07_util.R
\name{driesner07_H2O_NaCl_get_Tp_curve_on_F_VL_boundary}
\alias{driesner07_H2O_NaCl_get_Tp_curve_on_F_VL_boundary}
\title{get an array of T and p values along a phase boundary between F and VL regions}
\usage{
driesner07_H2O_NaCl_get_Tp_curve_on_F_VL_boundary(massfrac_NaCl, TCmax)
}
\arguments{
\item{massfrac_NaCl}{NaCl mass fraction (bulk) [kg/kg]}

\item{TCmax}{Maximum temperature to be calculated [K]}
}
\value{
dataframe object including calculated TK and p values
}
\description{
get an array of T and p values along a phase boundary between F and VL regions
}
