% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/driesner07_util.R
\name{driesner07_H2O_NaCl_get_Tp_curve_on_VL_VH_boundary}
\alias{driesner07_H2O_NaCl_get_Tp_curve_on_VL_VH_boundary}
\title{get an array of T and p values along a phase boundary between VL and VH regions}
\usage{
driesner07_H2O_NaCl_get_Tp_curve_on_VL_VH_boundary(TCmax, n = 50)
}
\arguments{
\item{TCmax}{Maximum temperature to be calculated [K]}

\item{n}{the number of sampling temperatures}
}
\value{
dataframe object including calculated TK and p values
}
\description{
get an array of T and p values along a phase boundary between VL and VH regions
}
