% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/driesner07.R
\name{driesner07_H2O_NaCl_LH_xl_Tp}
\alias{driesner07_H2O_NaCl_LH_xl_Tp}
\title{Saturated liquid composition at L+H coexisting, Function of Temperature and Pressure}
\usage{
driesner07_H2O_NaCl_LH_xl_Tp(TK, p)
}
\arguments{
\item{TK}{Temperature [K]}

\item{p}{Pressure [Pa]}
}
\value{
Mole fraction [-]
}
\description{
The function returns the saturated liquid composition at L+Halite coexisting surface
for given temperature and pressure.
}
