% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/driesner07.R
\name{driesner07_H2O_NaCl_pc_T}
\alias{driesner07_H2O_NaCl_pc_T}
\title{Critical pressuer of H2O-NaCl fluid as a function of temperature}
\usage{
driesner07_H2O_NaCl_pc_T(TK)
}
\arguments{
\item{TK}{Temperature [K]}
}
\value{
Critical pressure [Pa]
}
\description{
Returns critical pressure of H2O-NaCl fluid at the given temperature
}
