% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/driesner07_util.R
\name{driesner07_H2O_NaCl_phase_name_Tpx}
\alias{driesner07_H2O_NaCl_phase_name_Tpx}
\title{Phase name of H2O-NaCl fluid as a function of temperature, pressure, and NaCl composition}
\usage{
driesner07_H2O_NaCl_phase_name_Tpx(TK, p, x)
}
\arguments{
\item{TK}{Temeprature [K]}

\item{p}{Pressure [Pa]}

\item{x}{Bulk composition (mole fraction) of NaCl [mol/mol]}
}
\value{
phase name
}
\description{
Phase name of H2O-NaCl fluid as a function of temperature, pressure, and NaCl composition
}
