% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/driesner07.R
\name{driesner07_H2O_NaCl_phase_Tpx}
\alias{driesner07_H2O_NaCl_phase_Tpx}
\title{Phase id of H2O-NaCl fluid as a function of temperature, pressure, and NaCl composition}
\usage{
driesner07_H2O_NaCl_phase_Tpx(TK, p, x)
}
\arguments{
\item{TK}{Temeprature [K]}

\item{p}{Pressure [Pa]}

\item{x}{Bulk composition (mole fraction) of NaCl [mol/mol]}
}
\value{
phase id
}
\description{
Returns phase id of H2O-NaCl fluid at the given temperature, pressure, and bulk salinity
}
