% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/driesner07_util.R
\name{driesner07_H2O_NaCl_plot_phase_relation_on_Tp_space}
\alias{driesner07_H2O_NaCl_plot_phase_relation_on_Tp_space}
\title{plot H2O-NaCl phase diagram on T-p space}
\usage{
driesner07_H2O_NaCl_plot_phase_relation_on_Tp_space(dat, TC_range_max = 800)
}
\arguments{
\item{dat}{H2O-NaCl phase relation information}

\item{TC_range_max}{Maximum temperature to be plotted [K]}
}
\description{
plot H2O-NaCl phase diagram on T-p space
}
