\name{driesner07_H2O_NaCl_rho_pTx}
\alias{driesner07_H2O_NaCl_rho_pTx}
\title{H2O-NaCl fluid density model in Driesner (2007)}

\arguments{
\item{p_Pa}{pressure [Pa]}
\item{T_K}{temperature [K]}
\item{x}{NaCl mole composition [mol/mol]}
}
\value{
H2O-NaCl fluid density [kg/m^3]
}

\description{
Calculate H2O-NaCl fluid density from Driesner (2007)   
}
\references{
Driesner, T., Heinrich, C.A. (2007) The system H2O-NaCl. Part I: correlation formulae for 
phase relations in temperature-pressure-composition space from 0 to 1000 ◦c, 0 to 5000 bar, 
and 0 to 1 X_nacl, Geochim. Cosmochim. Acta 71 (20), 4880-4901.

Driesner, T. (2007) The system H2O-NaCl. part II: correlations for molar volume, 
enthalpy, and isobaric heat capacity from 0 to 1000 ◦C, 1 to 5000 bar, and 0 to 1 X_NaCl, 
Geochim. Cosmochim. Acta 71 (20), 4902-4919.
}

