% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/driesner07.R
\name{driesner07_H2O_NaCl_singlephase_h_Tpx}
\alias{driesner07_H2O_NaCl_singlephase_h_Tpx}
\title{Salinewater Specific Enthalphy, Function of Temperature, Pressure, and composition}
\usage{
driesner07_H2O_NaCl_singlephase_h_Tpx(TK, p, x)
}
\arguments{
\item{TK}{Temperature [K]}

\item{p}{Pressure [Pa]}

\item{x}{Mole fraction [-]}
}
\value{
Specific Enthalpy [J/kg]
}
\description{
The function returns the saline water specific enthalpy for given temperature, pressure,
and composition.
}
