% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/driesner07.R
\name{driesner07_H2O_NaCl_singlephase_p_rhoTx}
\alias{driesner07_H2O_NaCl_singlephase_p_rhoTx}
\title{calculate pressure from salinewater density and salinity}
\usage{
driesner07_H2O_NaCl_singlephase_p_rhoTx(rho, TK, x, pVL, pmax = 5e+08)
}
\arguments{
\item{rho}{Density [kg/m^3]}

\item{TK}{Temperature [K]}

\item{x}{Mole fraction [-]}

\item{pVL}{V-L pressure used as lower bound [Pa]}

\item{pmax}{upper bound of pressure [Pa]}
}
\value{
pressure [Pa]
}
\description{
calculate pressure from salinewater density and salinity
}
