% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/driesner07.R
\name{driesner07_H2O_NaCl_singlephase_rho_pTx}
\alias{driesner07_H2O_NaCl_singlephase_rho_pTx}
\title{Salinewater density, Function of Temperature, Pressure, and composition}
\usage{
driesner07_H2O_NaCl_singlephase_rho_pTx(p, TK, x)
}
\arguments{
\item{p}{Pressure [Pa]}

\item{TK}{Temperature [K]}

\item{x}{Mole fraction [-]}
}
\value{
Density [kg/m3]
}
\description{
The function returns the saline water density for given temperature, pressure,
and composition.
}
