% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/driesner07.R
\name{driesner07_H2O_NaCl_Tc_x}
\alias{driesner07_H2O_NaCl_Tc_x}
\title{Critical temperature of H2O-NaCl fluid as a function of NaCl composition}
\usage{
driesner07_H2O_NaCl_Tc_x(x)
}
\arguments{
\item{x}{Composition (mole fraction) of NaCl [mol/mol]}
}
\value{
Temperature [K]
}
\description{
Returns critical temperature of H2O-NaCl fluid at the given composition
}
