% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/driesner07.R
\name{driesner07_H2O_NaCl_VH_vol_frac_h}
\alias{driesner07_H2O_NaCl_VH_vol_frac_h}
\title{Halite volume fraction in V+H phase}
\usage{
driesner07_H2O_NaCl_VH_vol_frac_h(TK, p, x)
}
\arguments{
\item{TK}{Temperature [K]}

\item{p}{Pressure [Pa]}

\item{x}{NaCl composition [mol/mol]}
}
\value{
Volume fraction [m3/m3]
}
\description{
The function returns the solid volume fraction in V+H phase
for given temperature, pressure, and composition.
}
