% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/driesner07.R
\name{driesner07_H2O_NaCl_VH_xv_Tp}
\alias{driesner07_H2O_NaCl_VH_xv_Tp}
\title{Saturated vapor composition at V+H coexisting, Function of Temperature and Pressure}
\usage{
driesner07_H2O_NaCl_VH_xv_Tp(TK, p)
}
\arguments{
\item{TK}{Temperature [K]}

\item{p}{Pressure [Pa]}
}
\value{
Mole fraction [-]
}
\description{
The function returns the saturated vapor composition at V+Halite coexisting surface
for given temperature and pressure.
}
