% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/driesner07.R
\name{driesner07_H2O_NaCl_VL_mass_frac_l}
\alias{driesner07_H2O_NaCl_VL_mass_frac_l}
\title{Liquid mass fraction in V+L phase}
\usage{
driesner07_H2O_NaCl_VL_mass_frac_l(TK, p, x)
}
\arguments{
\item{TK}{Temperature [K]}

\item{p}{Pressure [Pa]}

\item{x}{NaCl composition [mol/mol]}
}
\value{
Mass fraction [kg/kg]
}
\description{
The function returns the liquid mass fraction in V+L phase
for given temperature, pressure, and composition.
}
