% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/driesner07_util.R
\name{driesner07_H2O_NaCl_VL_pv_Tx}
\alias{driesner07_H2O_NaCl_VL_pv_Tx}
\title{get vapor phase pressure(s) in V+L region at the given T and x}
\usage{
driesner07_H2O_NaCl_VL_pv_Tx(TK, x)
}
\arguments{
\item{TK}{Temperature [K]}

\item{x}{Bulk composition (mole fraction) of NaCl [mol/mol]}
}
\value{
a list of vapor pressures [Pa]
}
\description{
get vapor phase pressure(s) in V+L region at the given T and x
}
