% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/driesner07.R
\name{driesner07_H2O_NaCl_VL_rhov_Tp}
\alias{driesner07_H2O_NaCl_VL_rhov_Tp}
\title{Saturated vapor density at V+L surface, Function of Temperature and Pressure}
\usage{
driesner07_H2O_NaCl_VL_rhov_Tp(TK, p)
}
\arguments{
\item{TK}{Temperature [K]}

\item{p}{Pressure [Pa]}
}
\value{
Density [kg/m3]
}
\description{
The function returns the saturated vapor density at V+L surface
for given temperature and pressure.
}
