% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/driesner07.R
\name{driesner07_H2O_NaCl_VLH_p_T}
\alias{driesner07_H2O_NaCl_VLH_p_T}
\title{Saturation pressure at V+L+H surface, Function of Temperature}
\usage{
driesner07_H2O_NaCl_VLH_p_T(TK)
}
\arguments{
\item{TK}{Temperature [K]}
}
\value{
Pressure [Pa]
}
\description{
The function returns the saturation pressure at V+L+Halite coexisting surface
for given temperature.
}
