% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/driesner07_util.R
\name{driesner07_H2O_NaCl_VLH_T_xv}
\alias{driesner07_H2O_NaCl_VLH_T_xv}
\title{get temperature of VLH curve at the given vapor x}
\usage{
driesner07_H2O_NaCl_VLH_T_xv(xv)
}
\arguments{
\item{xv}{NaCl composition (mole fraction) of vapor [mol/mol]}
}
\value{
Temperature [K]
}
\description{
get temperature of VLH curve at the given vapor x
}
