% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/driesner07.R
\name{driesner07_H2O_NaCl_VLH_xl_T}
\alias{driesner07_H2O_NaCl_VLH_xl_T}
\title{Saturated liquid composition at V+L+H surface, Function of Temperature}
\usage{
driesner07_H2O_NaCl_VLH_xl_T(TK)
}
\arguments{
\item{TK}{Temperature [K]}
}
\value{
Mole fraction [-]
}
\description{
The function returns the saturated liquid composition at V+L+Halite coexisting surface
for given temperature.
}
