% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/driesner07.R
\name{driesner07_H2O_NaCl_xc_T}
\alias{driesner07_H2O_NaCl_xc_T}
\title{Critical NaCl composition of H2O-NaCl fluid as a function of temperature}
\usage{
driesner07_H2O_NaCl_xc_T(TK)
}
\arguments{
\item{TK}{Temperature [K]}
}
\value{
Critical composition (mole fraction) of NaCl [mol/mol]
}
\description{
Returns critical composition of H2O-NaCl fluid at the given temperature
}
