% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/driesner07.R
\name{driesner07_NaCl_H_cp_pT}
\alias{driesner07_NaCl_H_cp_pT}
\title{calculate halite specific heat}
\usage{
driesner07_NaCl_H_cp_pT(pPa, TK)
}
\arguments{
\item{pPa}{Pressure [Pa]}

\item{TK}{Temperature [K]}
}
\value{
specific heat [J/kg/K]
}
\description{
calculate halite specific heat
}
