% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/driesner07.R
\name{driesner07_NaCl_H_h_pT}
\alias{driesner07_NaCl_H_h_pT}
\title{calculate halite specific enthalpy}
\usage{
driesner07_NaCl_H_h_pT(pPa, TK)
}
\arguments{
\item{pPa}{Pressure [Pa]}

\item{TK}{Temperature [K]}
}
\value{
specific enthalpy [J/kg]
}
\description{
calculate halite specific enthalpy
}
