% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/driesner07.R
\name{driesner07_NaCl_H_rho_pT}
\alias{driesner07_NaCl_H_rho_pT}
\title{calculate halite density}
\usage{
driesner07_NaCl_H_rho_pT(pPa, TK)
}
\arguments{
\item{pPa}{Pressure [Pa]}

\item{TK}{Temperature [K]}
}
\value{
Density [kg/m^3]
}
\description{
calculate halite density
}
