% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/electrical_conductivity_rock.R
\name{granite_ec_Olhoeft1981_T}
\alias{granite_ec_Olhoeft1981_T}
\title{Electrical conductivity model for dry granite based on Olhoeft (1981, JGR) figure}
\usage{
granite_ec_Olhoeft1981_T(TK)
}
\arguments{
\item{TK}{Temperature [K]}
}
\value{
electrical conductivity of dry granite [S/m]
}
\description{
Electrical conductivity model for dry granite based on Olhoeft (1981, JGR) figure
}
