% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/H2ONaCl_utils.R
\name{H2ONaCl_b_to_M}
\alias{H2ONaCl_b_to_M}
\title{Convert H2O-NaCl concentration unit from molality to mole fraction}
\usage{
H2ONaCl_b_to_M(m, rho = 998.2)
}
\arguments{
\item{m}{molality [mol/kg-H2O]}

\item{rho}{density [kg/m3]}
}
\value{
Molar concentration [mol/L]
}
\description{
Convert concentration unit of saline water
}
