% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/H2ONaCl_utils.R
\name{H2ONaCl_b_to_massfrac}
\alias{H2ONaCl_b_to_massfrac}
\title{Convert H2O-NaCl concentration unit from molality to mass fraction}
\usage{
H2ONaCl_b_to_massfrac(b)
}
\arguments{
\item{b}{molality [mol/kg-H2O]}
}
\value{
mass fraction of NaCl [kg/kg]
}
\description{
Convert H2O-NaCl concentration unit from molality to mass fraction
}
