% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/H2ONaCl_utils.R
\name{H2ONaCl_b_to_x}
\alias{H2ONaCl_b_to_x}
\title{Convert H2O-NaCl concentration unit from molality to mole fraction}
\usage{
H2ONaCl_b_to_x(molality)
}
\arguments{
\item{molality}{molality [mol/kg-H2O]}
}
\value{
Mole fraction [-]
}
\description{
Convert concentration unit of saline water
}
