% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/H2ONaCl_electrical_conductivity.R
\name{H2ONaCl_ec_SakumaIchiki2016_highP}
\alias{H2ONaCl_ec_SakumaIchiki2016_highP}
\title{H2O-NaCl fluid electrical conductivity model in Sakuma and Ichiki (2016) for high pressures}
\usage{
H2ONaCl_ec_SakumaIchiki2016_highP(p, T, c)
}
\arguments{
\item{p}{Pressure [MPa]}

\item{T}{Temperature [K]}

\item{c}{Salinity [wt\%]}
}
\value{
Electrical conductivity [ohm^-1 m^-1]
}
\description{
The model is valid for 0.2-2 GPa, 673-2000K, 0.6-9.6wt\%.
}
\references{
Sakuma, H., Ichiki, M. (2016) Electrical conductivity of NaCl-H2O fluid
in the crust. JGR Solid Earth.
}
