% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/H2ONaCl_electrical_conductivity.R
\name{H2ONaCl_ec_SenGoode1992}
\alias{H2ONaCl_ec_SenGoode1992}
\title{H2O-NaCl fluid electrical conductivity model by Sen & Goode (1992)}
\usage{
H2ONaCl_ec_SenGoode1992(m, TC)
}
\arguments{
\item{m}{NaCl molality [mol/kg]}

\item{TC}{Temperature [deg C]}
}
\value{
Electrical conductivity [S/m]
}
\description{
The model is valid for 20-200 C.
}
\references{
Sen, P.N., Goode, P.A. (1992) Influence of temperature on electrical
conductivity of shaly sands, Geophysics 57 (1), 89–96.

Sen, P.N., Goode, P.A. (1992) Errata, to “influence of temperature on
electrical conduc- tivity of shaly sands”, Geophysics 57 (12), 1658.
}
