% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/H2ONaCl_electrical_conductivity.R
\name{H2ONaCl_ec_SinmyoKeppler2017}
\alias{H2ONaCl_ec_SinmyoKeppler2017}
\title{H2O-NaCl fluid electrical conductivity model by Sinmyo and Keppler (2017)}
\usage{
H2ONaCl_ec_SinmyoKeppler2017(pMPa, TK, c_wtp)
}
\arguments{
\item{pMPa}{Pressure [MPa]}

\item{TK}{Temperature [K]}

\item{c_wtp}{Salinity [wt\%]}
}
\value{
Electrical conductivity [ohm^-1 m^-1]
}
\description{
H2O-NaCl fluid electrical conductivity model by Sinmyo and Keppler (2017)
}
\references{
Sinmyo, R., Keppler, H. (2017) Electrical conductivity of NaCl-bearing aqueous
fluids to 600C and 1GPa. Contrib Mineral Petrol 172:4
}
