% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/H2ONaCl_electrical_conductivity.R
\name{H2ONaCl_ec_WatanabeEtAl2022_density_model_Tpm}
\alias{H2ONaCl_ec_WatanabeEtAl2022_density_model_Tpm}
\title{H2O-NaCl fluid electrical conductivity model for low density conditions in Watanabe et al (2022, Geothermics)}
\usage{
H2ONaCl_ec_WatanabeEtAl2022_density_model_Tpm(TK, pMPa, m)
}
\arguments{
\item{TK}{Temperature [K]}

\item{pMPa}{Pressure [MPa]}

\item{m}{NaCl molality [mol/kg-H2O]}
}
\value{
Electrical conductivity [S/m]
}
\description{
H2O-NaCl fluid electrical conductivity model for low density conditions in Watanabe et al (2022, Geothermics)
}
