% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/H2ONaCl_utils.R
\name{H2ONaCl_M_to_massfrac}
\alias{H2ONaCl_M_to_massfrac}
\title{Convert H2O-NaCl concentration unit from molar concentration to mass fraction}
\usage{
H2ONaCl_M_to_massfrac(M, rho = 998.2)
}
\arguments{
\item{M}{Molar concentraion [mol/L] = [M]}

\item{rho}{Solution density [kg/m3]}
}
\value{
Mass fraction [kg/kg]
}
\description{
Convert concentration unit of saline water from molar concentration to mass fraction
}
