% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/H2ONaCl_utils.R
\name{H2ONaCl_M_to_x}
\alias{H2ONaCl_M_to_x}
\title{Convert H2O-NaCl concentration unit from molar concentration to molar fraction}
\usage{
H2ONaCl_M_to_x(M, rho = 998.2)
}
\arguments{
\item{M}{Molarity [mol/L]}

\item{rho}{Solution density [kg/m3]}
}
\value{
Mole fraction of NaCl [mol/mol]
}
\description{
Convert H2O-NaCl concentration unit from molar concentration to molar fraction
}
