% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/H2ONaCl_utils.R
\name{H2ONaCl_massfrac_to_b}
\alias{H2ONaCl_massfrac_to_b}
\title{Convert H2O-NaCl concentration unit from mass fraction to molality}
\usage{
H2ONaCl_massfrac_to_b(X)
}
\arguments{
\item{X}{Mass fraction [kg/kg]}
}
\value{
molality [mol/kg-H2O]
}
\description{
Convert H2O-NaCl concentration unit from mass fraction to molality
}
