% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/H2ONaCl_utils.R
\name{H2ONaCl_massfrac_to_M}
\alias{H2ONaCl_massfrac_to_M}
\title{Convert H2O-NaCl concentration unit from mass fraction to molar concentration}
\usage{
H2ONaCl_massfrac_to_M(X, rho = 998.2)
}
\arguments{
\item{X}{Mass fraction [kg/kg]}

\item{rho}{Solution density [kg/m3]}
}
\value{
Molar concentraion [mol/L]
}
\description{
Convert H2O-NaCl concentration unit from mass fraction to molar concentration
}
