% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/H2ONaCl_utils.R
\name{H2ONaCl_massfrac_to_x}
\alias{H2ONaCl_massfrac_to_x}
\title{Convert H2O-NaCl concentration unit from mass fraction to mole fraction}
\usage{
H2ONaCl_massfrac_to_x(mass_frac)
}
\arguments{
\item{mass_frac}{Mass fraction [-]}
}
\value{
Mole fraction [-]
}
\description{
Convert concentration unit of saline water
}
