% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/H2ONaCl_electrical_conductivity.R
\name{H2ONaCl_molar_ec_WatanabeEtAl2021_vism}
\alias{H2ONaCl_molar_ec_WatanabeEtAl2021_vism}
\title{Molar electrical conductivity model of H2O-NaCl fluids in Watanabe et al (2021, FPE)}
\usage{
H2ONaCl_molar_ec_WatanabeEtAl2021_vism(vis, m)
}
\arguments{
\item{vis}{viscosity [Pa s]}

\item{m}{NaCl molality [mol/kg-H2O]}
}
\value{
Molar conductivity [Sm^2/mol]
}
\description{
Molar electrical conductivity model of H2O-NaCl fluids in Watanabe et al (2021, FPE)
}
