% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/H2ONaCl_utils.R
\name{H2ONaCl_x_to_massfrac}
\alias{H2ONaCl_x_to_massfrac}
\title{Convert H2O-NaCl concentration unit from mole fraction to mass fraction}
\usage{
H2ONaCl_x_to_massfrac(mole_frac)
}
\arguments{
\item{mole_frac}{Mole fraction [-]}
}
\value{
Mass fraction [-]
}
\description{
Convert concentration unit of saline water
}
