% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iapws95.R
\name{iapws95_rho_pT}
\alias{iapws95_rho_pT}
\title{Water density using IAPWS-95}
\usage{
iapws95_rho_pT(pPa, TK)
}
\arguments{
\item{pPa}{pressure [Pa]}

\item{TK}{Temperature [K]}
}
\value{
water density [kg/m^3]
}
\description{
Water density using IAPWS-95
}
