% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/klyukinetal17.R
\name{klyukinetal2017_H2O_NaCl_viscosity_rhoTx}
\alias{klyukinetal2017_H2O_NaCl_viscosity_rhoTx}
\title{Dynamic Viscosity of Salinewater, Function of Density, Temperature, and Salinity}
\usage{
klyukinetal2017_H2O_NaCl_viscosity_rhoTx(rho, TK, x)
}
\arguments{
\item{rho}{Fluid density [kg/m3]}

\item{TK}{Temperature [K]}

\item{x}{Salinity in mole fraction [mol/mol]}
}
\value{
Dynamic viscosity [Pa s]
}
\description{
Dynamic Viscosity of Salinewater, Function of Density, Temperature, and Salinity
}
\references{
Klyukin, Y.I., Lowell R.P., Bodnar, R.J. (2017) A revised empirical
model to calculate the dynamic viscosity of H2O-NaCl fluids at elevated
temperatures and pressures (<=1000 C, <=500 MPa, 0-100 wt % NaCl). Fluid Phase
Equilibria 433, 193-205.
}
