\name{klyukinetal2017_viscosity}
\alias{klyukinetal2017_viscosity}
\title{H2O-NaCl fluid viscosity in Klyukin et al. (2017)}

\arguments{
\item{rho}{H2O-NaCl fluid density [kg/m^3]}
\item{T}{temperature [K]}
\item{xNaCl}{NaCl mole composition [mol/mol]}
}
\value{
H2O-NaCl fluid viscosity [Pa s]
}

\description{
Calculate H2O-NaCl fluid viscosity from Klyukin et al. (2017)   
}
\references{
Klyukin, Y.I.., Lowell, R.P., Bodnar, R.J. (2017) A revised empirical model 
to calculate the dynamic viscosity of H2O-NaCl fluids at elevated temperatures 
and pressures (≤10 0 0 ◦c, ≤50 0 MPa, 0-10 0 wt \% NaCl), Fluid Phase Equilibria 433, 193-205
}
