% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/electrical_conductivity_melt.R
\name{melt_ec_rhyolite_Gaillard2004_Tpw}
\alias{melt_ec_rhyolite_Gaillard2004_Tpw}
\title{Electrical conductivity model for hydrous rhyolitic–granitic melts in Gaillard (2004, EPSL)}
\usage{
melt_ec_rhyolite_Gaillard2004_Tpw(TK, pMPa, wtp_H2O)
}
\arguments{
\item{TK}{Temperature [K]}

\item{pMPa}{Pressure [MPa]}

\item{wtp_H2O}{Weight percent of H2O [\%]}
}
\value{
electrical conductivity [S/m]
}
\description{
Electrical conductivity model for hydrous rhyolitic–granitic melts in Gaillard (2004, EPSL)
}
