% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/electrical_conductivity_melt.R
\name{melt_ec_rhyolite_GuoEtAl2016_Tpw}
\alias{melt_ec_rhyolite_GuoEtAl2016_Tpw}
\title{Electrical conductivity model for rhyolitic melt in Guo et al (2016, EPSL)}
\usage{
melt_ec_rhyolite_GuoEtAl2016_Tpw(TK, pGPa, wtp_H2O)
}
\arguments{
\item{TK}{Temperature [K]}

\item{pGPa}{Pressure [GPa]}

\item{wtp_H2O}{Weight percent of H2O [\%]}
}
\value{
electrical conductivity [S/m]
}
\description{
Electrical conductivity model for rhyolitic melt in Guo et al (2016, EPSL)
}
