% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/electrical_conductivity_bulk_model.R
\name{phi_Archie_ec}
\alias{phi_Archie_ec}
\title{Calculate porosity from bulk electrical conductivity using Archie's law}
\usage{
phi_Archie_ec(sigma_bulk, sigma_f, m)
}
\arguments{
\item{sigma_bulk}{bulk electrical conductivity [S/m]}

\item{sigma_f}{fluid electrical conductivity [S/m]}

\item{m}{cementation exponent [-]}
}
\value{
porosity [-]
}
\description{
Calculate porosity from bulk electrical conductivity using Archie's law
}
