% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/electrical_conductivity_bulk_model.R
\name{phi_HSupper_ec}
\alias{phi_HSupper_ec}
\title{Calculate porosity from bulk electrical conductivity using Hashin-Shtrikman upper bound}
\usage{
phi_HSupper_ec(sb, sf, ss)
}
\arguments{
\item{sb}{bulk electrical conductivity [S/m]}

\item{sf}{fluid electrical conductivity [S/m]}

\item{ss}{Solid electrical conductivity [S/m]}
}
\value{
porosity [-]
}
\description{
Calculate porosity from bulk electrical conductivity using Hashin-Shtrikman upper bound
}
