% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geofluidprop_common.R
\name{toC}
\alias{toC}
\title{convert temperature in Kelvin to degree C}
\usage{
toC(TK)
}
\arguments{
\item{TK}{temperature in K}
}
\value{
temperature in deg C
}
\description{
convert temperature in Kelvin to degree C
}
