% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geofluidprop_common.R
\name{toK}
\alias{toK}
\title{convert temperature in degree C to Kelvin}
\usage{
toK(TC)
}
\arguments{
\item{TC}{temperature in deg C}
}
\value{
temperature in K
}
\description{
convert temperature in degree C to Kelvin
}
